requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.22
numpy>=1.26.0
psycopg2-binary>=2.9.9
chromadb>=0.4.18
python-dotenv>=1.0.0
